Move external drag support to androidx.compose.ui package
