+++ compose/desktop/desktop/samples/src/jvmMain/kotlin/androidx/compose/desktop/examples/dnd/Main.jvm.kt
@@ -29,12 +29,12 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.DropData
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.dnd.DropData
-import androidx.compose.ui.dnd.onExternalDrag
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.onExternalDrag
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.singleWindowApplication
